- Reworked ObjectExpression (and all Object*) to remove requirement for all function and mediator implementations to have to manage these types (#286)
- Added user defined clr types to support scenarios where column data contains the representation of a more complex object; i.e. JSON, XML, binary serialized data
- Reworked functions (only one required was Coalesce) to return a typed element when provided a generic type when fluently building the function (i.e. db.fx.Coalesce<int>(...) results in AnyElement<int> and db.fx.Coalesce(...) results in ObjectElement)
- Reworked interfaces for *Elements to have a more correct inheritance hierarchy
- Reworked fluent entry to select one and select many to accept more abstract type (i.e. AnyElement<bool> instead of BooleanElement)
- Removed all typed select expressions in favor of generic select expression (i.e. SelectExpression<bool> instead of BooleanSelectExpression)
- Hide TableAlias and FieldAlias properties from fluent building of query expressions
- Reworked aliasing to return generic select expression (i.e. SelectExpression<bool> instead of BooleanSelectExpression)

+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/_Functions/PopulationVariance.cs
@@ -131,7 +131,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "SUBQUERY")]
-        public void Can_populationvariance_of_aliased_field_succeed(int version, double expected = 18.6667)
+        public void Can_populationvariance_of_aliased_field_succeed(int version, float expected = 18.6667f)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -148,7 +148,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
             object result = exp.Execute();
 
             //then
-            result.Should().BeOfType<double>().Which.Should().BeApproximately(expected, 0.0001, "Rounding errors in population standard deviation");
+            result.Should().BeOfType<float>().Which.Should().BeApproximately(expected, 0.0001f, "Rounding errors in population standard deviation");
         }
     }
 }
